Move getXInterval to renderer_utils.js and use it for X value for waterfall total
